GUI refactoring. All Tested
