+++ operatoractivitiesmanager/src/main/java/com/salvatorecerullo/app/operatoractivitiesmanager/view/swing/OperatorActivitiesManagerView.java
@@ -11,16 +11,14 @@ import javax.swing.WindowConstants;
 import javax.swing.BoxLayout;
 
 public class OperatorActivitiesManagerView extends JFrame {
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = -4591844476430172042L;
+
 	private JPanel contentPane;
+	private JTabbedPane tabbedPane;
+
 	private ActivitiesPanel activitiesPanel;
 	private OperatorsPanel operatorsPanel;
 	private BasicOperationPanel basicOperationPanel;
-
-	private JTabbedPane tabbedPane;
 
 	/**
 	 * Launch the application.
@@ -30,32 +28,37 @@ public class OperatorActivitiesManagerView extends JFrame {
 	 * new OperatorActivitiesManagerView(); frame.setVisible(true); } catch
 	 * (Exception e) { e.printStackTrace(); } } }); }
 	 */
-	/**
-	 * Create the frame.
-	 */
+
 	public OperatorActivitiesManagerView() {
 		setTitle("Operator Activities Manager");
 		setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
 		setBounds(100, 100, 794, 734);
+		
+		// CONTENT PANE
 		contentPane = new JPanel();
 		contentPane.setBorder(new EmptyBorder(5, 5, 5, 5));
 		setContentPane(contentPane);
 		contentPane.setName("contentPane");
 		contentPane.setLayout(new BoxLayout(contentPane, BoxLayout.X_AXIS));
 
+		// TABBED PANE
 		tabbedPane = new JTabbedPane(SwingConstants.TOP);
 		contentPane.add(tabbedPane);
 		tabbedPane.setName("tabbedPane");
 
+		// ACTIVITIES PANEL
 		activitiesPanel = new ActivitiesPanel();
 		tabbedPane.addTab("Activities", null, activitiesPanel, null);
+
+		// OPERATORS PANEL
 		operatorsPanel = new OperatorsPanel();
 		tabbedPane.addTab("Operators", null, operatorsPanel, null);
+
+		// BASIC OPERATION PANEL
 		basicOperationPanel = new BasicOperationPanel();
 		tabbedPane.addTab("Basic Operations", null, basicOperationPanel, null);
 
 		tabbedPane.addChangeListener(e -> changeListenerTabbedPane());
-
 		setNames();
 	}
